Extract CopyTailwinBuildPlugin
