+++ test-projects/scenario-4-addon-using-sass/sample-addon/index.js
@@ -1,9 +1,7 @@
 'use strict';
 
-const path = require('path');
-const BroccoliPlugin = require('broccoli-plugin');
 const Merge = require('broccoli-merge-trees');
-const fs = require('fs-extra');
+const CopyTailwindBuildPlugin = require('ember-cli-tailwind/lib/copy-tailwind-build-plugin');
 
 module.exports = {
   name: 'sample-addon',
@@ -19,26 +17,8 @@ module.exports = {
   treeForAddonStyles(tree) {
     let trees = tree ? [ tree ] : [];
 
-    trees.push(new AddFilePlugin([ tree ], this));
+    trees.push(new CopyTailwindBuildPlugin([ tree ], this));
 
     return new Merge(trees);
   }
 };
-
-
-class AddFilePlugin extends BroccoliPlugin {
-
-  constructor(inputTrees, project) {
-    super(inputTrees);
-
-    this.project = project;
-  }
-
-  build() {
-    let outputFile = path.join(this.outputPath, 'tailwind.css');
-    let tailwindFile = this.project.findOwnAddonByName('ember-cli-tailwind').tailwindOutputFile;
-    let tailwindCSS = fs.readFileSync(tailwindFile, 'utf-8');
-
-    fs.writeFileSync(outputFile, tailwindCSS);
-  }
-}
